pyproject.toml
[tool.poetry]
name = "endopulse"
version = "0.1.0"
description = ""
authors = ["jon <jonathan.n@u.nus.edu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-telegram-bot = {extras = ["job-queue"], version = "^20.6"}
schedule = "^1.2.1"
pytz = "^2023.3.post1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
